+++ src/app/services/character.service.mock.ts
@@ -115,26 +115,25 @@ export const getMockCharacter = (): Character=>   {
   pj.setTitles(['Regateador (10)', 'Temido (7)', 'Intuitivo (6)', 'Orador (7)']);
 
   pj.setOtherNotes(['Hijo de Hermes', 'Capa mágica: Saltar al agua transporta al grupo junto a mi barco.']);
-  let history = 'Mi nombre es Dymas de Megalopolis, naci hace 180 inviernos en el seno de una famila modesta, mi madre, que aún vivia con mis abuelos ';
-  history = history.concat(' en una pequeña familia de pastores nunca dió explicaciones de como pasó, pero fué seducida por Hermes y algún tiempo después nací yo. Los primeros años de ');
-  history = history.concat('mi vida fueron quizá los más felices ');
-  history = history.concat('Mi juventud en Megalópolis fue una épocade excesos y corredurias poco épicas. ');
-  history = history.concat('Durante mi vigesimo octavo invierno, Hermes mi padre de habló por primera vez, prefiero no reproducir aquellas palabras, no fueron muy amables, me echó en cara mi modo de vida y me pidió un favor ');
-  history = history.concat(' otro hijo suyo, aun muy joven, habia caido en manos peligrosas y quería que yo lo salvase, me negé en redondo, cosa de la que no estoy orgulloso y cuando insistió le dije que no estaba dispuesto a perder los ');
-  history = history.concat(' años de mi vida huyendo con un niño ');
-  history = history.concat(' Un par de semanas después vi al niño en un mercado de esclavos junto al rio Helisonte, su mirada fué devastadora, tocó algo en mi interior. Esa noche me escurrí sin ser visto entre los esclavos y saqué al niño ');
-  history = history.concat(' a nado por el rio hasta una pequeña barca con la que huimos de la ciudad. Mi hermano me preguntó que como podía ser tan necio de haber accedido a los deseos de Hermes tras primero haberlos negado y que si era ');
-  history = history.concat('consciente de que padre no nos ayudaría en esas condiciones, por primera vez nada me importaba, durante unas semanas navegamos rio abajo y nos hicimos pasar por hermanos que viajabamos para comerciar con los pueblos pescadores de la costa');
-  history = history.concat(' Mas pronto que tarde descubrí que podía dejar de fingir y comerciar realmente, ganando unas monedas con las que comprar algo que echarnos a la boca ');
-  history = history.concat(' Pasaron los años, mi hermano creció pero no así yo, por algún motivo padre me permitió disfrutar de mi juventud mientras estuviese cuidando de su prole...');
-  history = history.concat(' Años despues encontré un niño medio ahogado flotando en una tabla junto al barco que ya poseía y de inmedia supe lo que habia que hacer, rescatéal niño y lo cuidé entre mis mejores pieles');
-  history = history.concat(' Ese niño resultó haber caido al agua desde un barco de alguien importante en Megalópolis, y los guardias lo reconocieron en mi barco, pensando que lo habia secuestrado... otra vez tuve que huir.');
-  history = history.concat(' Por alguna razón este niño estaba encantado de la vida en el barco, hasta talpunto que se convirtió en un par de años en un gran marinero');
-  history = history.concat(' Hasta que llegó un día en el que le pudo la responsabilidad y volvió con los suyos, no voy a hablar de quien era, pero más de quien yo creí...');
-  history = history.concat(' Unos 20 años más tarde mi fortuna habia crecido, y mi barco... Apareció un hombre alto y recio que me buscaba, era mi ex-marinero me pedia pasar una temporada viajando y aprendiendo de la mar, como hizo su padre antes que el');
-  history = history.concat(' Pasaron los años y la historia se repitió muchas veces más a veces no eran uno, si no doso tres, finalmente los lazos de la familia seperpetuaron de alguna forma conmigo y me convertí en el maestro de muchos de ellos...');
-  history = history.concat(' Aun en día un familiar lejado de dicha familia viaja junto a mi barco, ocupanto el puesto de mi primero de abordo y siendo alguien importante parami.');
-  history = history.concat(' Los años me han convertido en un gran comerciante y embaucador, yme han dado conocimientos en muchas areas, si bien quizá me vence en ciertos momentos el peso del oro...');
+  let history =  `Mi nombre es Dymas de Megalopolis, naci hace 180 inviernos en el seno de una famila modesta, mi madre, que aún vivia con mis abuelos
+ en una pequeña familia de pastores nunca dió explicaciones de como pasó, pero fué seducida por Hermes y algún tiempo después nací yo. Los primeros años de mi vida fueron quizá los más felices
+ Mi juventud en Megalópolis fue una épocade excesos y corredurias poco épicas.
+ Durante mi vigesimo octavo invierno, Hermes mi padre de habló por primera vez, prefiero no reproducir aquellas palabras, no fueron muy amables, me echó en cara mi modo de vida y me pidió un favor
+  otro hijo suyo, aun muy joven, habia caido en manos peligrosas y quería que yo lo salvase, me negé en redondo, cosa de la que no estoy orgulloso y cuando insistió le dije que no estaba dispuesto
+  a perder los años de mi vida huyendo con un niño. Un par de semanas después vi al niño en un mercado de esclavos junto al rio Helisonte, su mirada fué devastadora, tocó algo en mi interior.
+  Esa noche me escurrí sin ser visto entre los esclavos y saqué al niño a nado por el rio hasta una pequeña barca con la que huimos de la ciudad. Mi hermano me preguntó que como podía ser tan necio
+  de haber accedido a los deseos de Hermes tras primero haberlos negado y que si era consciente de que padre no nos ayudaría en esas condiciones, por primera vez nada me importaba, durante unas
+  semanas navegamos rio abajo y nos hicimos pasar por hermanos que viajabamos para comerciar con los pueblos pescadores de la costa Mas pronto que tarde descubrí que podía dejar de fingir y
+  comerciar realmente, ganando unas monedas con las que comprar algo que echarnos a la boca. Pasaron los años, mi hermano creció pero no así yo, por algún motivo padre me permitió disfrutar de mi
+  juventud mientras estuviese cuidando de su prole... Años despues encontré un niño medio ahogado flotando en una tabla junto al barco que ya poseía y de inmedia supe lo que habia que hacer, rescaté
+  al niño y lo cuidé entre mis mejores pieles. Ese niño resultó haber caido al agua desde un barco de alguien importante en Megalópolis, y los guardias lo reconocieron en mi barco, pensando que lo
+  habia secuestrado... otra vez tuve que huir. Por alguna razón este niño estaba encantado de la vida en el barco, hasta talpunto que se convirtió en un par de años en un gran marinero'
+  Hasta que llegó un día en el que le pudo la responsabilidad y volvió con los suyos, no voy a hablar de quien era, pero más de quien yo creí... Unos 20 años más tarde mi fortuna habia crecido,
+  y mi barco... Apareció un hombre alto y recio que me buscaba, era mi ex-marinero me pedia pasar una temporada viajando y aprendiendo de la mar, como hizo su padre antes que el. Pasaron los años y
+  la historia se repitió muchas veces,  más a veces no eran uno, si no doso tres, finalmente los lazos de la familia se perpetuaron de alguna forma conmigo y me convertí en el maestro de muchos de
+  ellos...Aun en día un familiar lejado de dicha familia viaja junto a mi barco, ocupanto el puesto de mi primero de abordo y siendo alguien importante parami. Los años me han convertido en un gran
+  comerciante y embaucador, yme han dado conocimientos en muchas areas, si bien quizá me vence en ciertos momentos el peso del oro...`;
+
   pj.setHistory(history);
 
   return pj;